use Time field of ArrivalTime as eta
CO-998

instead of calculating the eta from the Seconds field

this is more stable. it seems the response to GetStopArrivalTimes can be
cached for up to 5 seconds or so; computing the eta from the relative
Seconds field can lead to the eta steadily increasing for about five
seconds after the first request. using the Time field leads to a stable
eta.