add health check??
CO-998

Issue an API call every so often to make sure that the RideSystems API
is still running. This is probably kind of expensive, but they don't
seem to offer any sort of Ping endpoint, so. Good thing we never
actually run health checks on any of our APIs, so there's no way this
can cause a performance problem.