add a ResourceMapper test
CO-998

Add a simple test for ResourceMapper.mapRoute that just checks the
result on a hardcoded input. This doesn't test quite as much as i would
like: the input to mapRoute is a java object; i would like to also test
that the json response from RideSystems is parsed correctly, but that's
a little harder.